final_showcase/requirements.txt
pygame-ce==2.5.2
pygame_gui==0.6.12
openai==0.28.0
